feat: enhance knowledge API models and examples
- Add new model classes for data source, document metadata, external knowledge
- Enhance existing models with better type definitions and validation
- Update all knowledge API examples with improved error handling
- Refactor model structure for better maintainability
- Update documentation and tests
- Remove obsolete test files

+++ dify_oapi/api/knowledge/v1/model/file_info.py
@@ -17,7 +17,7 @@ class FileInfo(BaseModel):
     mime_type: Optional[str] = None
     type: Optional[FileType] = None
     created_by: Optional[str] = None
-    created_at: Optional[int] = None
+    created_at: Optional[float] = None
     upload_file_id: Optional[str] = None
 
     @staticmethod
@@ -62,7 +62,7 @@ class FileInfoBuilder:
         self._file_info.created_by = created_by
         return self
 
-    def created_at(self, created_at: int) -> "FileInfoBuilder":
+    def created_at(self, created_at: float) -> "FileInfoBuilder":
         self._file_info.created_at = created_at
         return self
