+++ dify_oapi/api/knowledge/v1/model/segment_info.py
@@ -22,15 +22,15 @@ class SegmentInfo(BaseModel):
     index_node_hash: Optional[str] = None
     hit_count: Optional[int] = None
     enabled: Optional[bool] = None
-    disabled_at: Optional[int] = None
+    disabled_at: Optional[float] = None
     disabled_by: Optional[str] = None
     status: Optional[SegmentStatus] = None
     created_by: Optional[str] = None
-    created_at: Optional[int] = None
-    indexing_at: Optional[int] = None
-    completed_at: Optional[int] = None
+    created_at: Optional[float] = None
+    indexing_at: Optional[float] = None
+    completed_at: Optional[float] = None
     error: Optional[str] = None
-    stopped_at: Optional[int] = None
+    stopped_at: Optional[float] = None
 
     @staticmethod
     def builder() -> "SegmentInfoBuilder":
@@ -94,7 +94,7 @@ class SegmentInfoBuilder:
         self._segment_info.enabled = enabled
         return self
 
-    def disabled_at(self, disabled_at: int) -> "SegmentInfoBuilder":
+    def disabled_at(self, disabled_at: float) -> "SegmentInfoBuilder":
         self._segment_info.disabled_at = disabled_at
         return self
 
@@ -110,15 +110,15 @@ class SegmentInfoBuilder:
         self._segment_info.created_by = created_by
         return self
 
-    def created_at(self, created_at: int) -> "SegmentInfoBuilder":
+    def created_at(self, created_at: float) -> "SegmentInfoBuilder":
         self._segment_info.created_at = created_at
         return self
 
-    def indexing_at(self, indexing_at: int) -> "SegmentInfoBuilder":
+    def indexing_at(self, indexing_at: float) -> "SegmentInfoBuilder":
         self._segment_info.indexing_at = indexing_at
         return self
 
-    def completed_at(self, completed_at: int) -> "SegmentInfoBuilder":
+    def completed_at(self, completed_at: float) -> "SegmentInfoBuilder":
         self._segment_info.completed_at = completed_at
         return self
 
@@ -126,6 +126,6 @@ class SegmentInfoBuilder:
         self._segment_info.error = error
         return self
 
-    def stopped_at(self, stopped_at: int) -> "SegmentInfoBuilder":
+    def stopped_at(self, stopped_at: float) -> "SegmentInfoBuilder":
         self._segment_info.stopped_at = stopped_at
         return self
